pyproject.toml
[project]
name = "llm-decision-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "streamlit>=1.52.2",
]
